+++ dev-test/general/githunt/graphql-declared-modules.d.ts
@@ -1,12 +1,3 @@
-declare module '*/current-user.query.graphql' {
-  import { DocumentNode } from 'graphql';
-  const defaultDocument: DocumentNode;
-  export const CurrentUserForProfileFromOutsideDirectory: DocumentNode;
-  export const CurrentUserForProfile: DocumentNode;
-
-  export default defaultDocument;
-}
-
 declare module '*/comment-added.subscription.graphql' {
   import { DocumentNode } from 'graphql';
   const defaultDocument: DocumentNode;
@@ -27,6 +18,14 @@ declare module '*/comments-page-comment.fragment.graphql' {
   import { DocumentNode } from 'graphql';
   const defaultDocument: DocumentNode;
   export const CommentsPageComment: DocumentNode;
+
+  export default defaultDocument;
+}
+
+declare module '*/current-user.query.graphql' {
+  import { DocumentNode } from 'graphql';
+  const defaultDocument: DocumentNode;
+  export const CurrentUserForProfile: DocumentNode;
 
   export default defaultDocument;
 }